refactor: rename backend module to runner
