+++ src/orquestra/integrations/qiskit/runner/__init__.py
@@ -3,3 +3,4 @@
 ################################################################################
 from ._qiskit_runner import QiskitRunner
 from ._qiskit_wavefunction_simulator import QiskitWavefunctionSimulator
+from ._ibmq_runner import create_ibmq_runner
