+++ tests/orquestra/integrations/qiskit/runner/_qiskit_wavefunction_simulator_test.py
@@ -1,7 +1,7 @@
 import pytest
 from qiskit import Aer
 
-from orquestra.integrations.qiskit.backend import QiskitWavefunctionSimulator
+from orquestra.integrations.qiskit.runner import QiskitWavefunctionSimulator
 from orquestra.quantum.api.circuit_runner_contracts import CIRCUIT_RUNNER_CONTRACTS
 from orquestra.quantum.api.wavefunction_simulator_contracts import simulator_contracts_for_tolerance
 from orquestra.quantum.circuits import Circuit, X, CNOT
